+++ src/invariants/XtYK.sol
@@ -5,17 +5,17 @@ import {ICurve, Pair} from "../interfaces/ICurve.sol";
 import "forge-std/Test.sol";
 
 contract XtYK is ICurve {
-    uint96 public constant MIN_LP = 10**3;
+    uint128 public constant MIN_LP = 10**3;
 
     function addLiquidity(
         Pair memory pair,
-        uint96 token0Amount,
-        uint96 token1Amount
-    ) external pure returns (uint96 k) {
-        uint96 reserve0 = pair.reserve0;
-        uint96 reserve1 = pair.reserve1;
+        uint128 token0Amount,
+        uint128 token1Amount
+    ) external pure returns (uint128 k) {
+        uint128 reserve0 = pair.reserve0;
+        uint128 reserve1 = pair.reserve1;
 
-        uint96 totalK = reserve0 * reserve1;
+        uint128 totalK = reserve0 * reserve1;
 
         if (totalK == 0) {
             k = token0Amount * token1Amount - MIN_LP;
@@ -26,15 +26,15 @@ contract XtYK is ICurve {
         require(k != 0, "insufficient K");
     }
 
-    function removeLiquidity(Pair memory pair, uint96 k)
+    function removeLiquidity(Pair memory pair, uint128 k)
         external
         pure
-        returns (uint96 amount0Out, uint96 amount1Out)
+        returns (uint128 amount0Out, uint128 amount1Out)
     {
-        uint96 reserve0 = pair.reserve0;
-        uint96 reserve1 = pair.reserve1;
+        uint128 reserve0 = pair.reserve0;
+        uint128 reserve1 = pair.reserve1;
 
-        uint96 totalK = reserve0 * reserve1;
+        uint128 totalK = reserve0 * reserve1;
 
         amount0Out = (k * reserve0) / totalK;
         amount1Out = (k * reserve1) / totalK;
@@ -43,22 +43,22 @@ contract XtYK is ICurve {
     }
 
     function swap(
-        uint96 reserveIn,
-        uint96 reserveOut,
-        uint96 amountIn
-    ) external pure returns (uint96 amountOut) {
+        uint128 reserveIn,
+        uint128 reserveOut,
+        uint128 amountIn
+    ) external pure returns (uint128 amountOut) {
         amountOut = _getAmountOut(amountIn, reserveIn, reserveOut);
     }
 
     function _getAmountOut(
-        uint96 amountIn,
-        uint96 reserveAmountIn,
-        uint96 reserveAmountOut
-    ) internal pure returns (uint96 amountOut) {
+        uint128 amountIn,
+        uint128 reserveAmountIn,
+        uint128 reserveAmountOut
+    ) internal pure returns (uint128 amountOut) {
         amountOut = (amountIn * reserveAmountOut) / reserveAmountIn;
     }
 
-    function _min(uint96 x, uint96 y) private pure returns (uint96 z) {
+    function _min(uint128 x, uint128 y) private pure returns (uint128 z) {
         z = x < y ? x : y;
     }
 }
